Fix a bug on the reference of the new ranges which was preventing the deletion [Redmine Issue #10618]
git-svn-id: http://euclid.esac.esa.int/svn/EC/SGS/SDC/CH/Projects/Phosphoros/trunk@13220 463500fe-d6be-4184-9b20-47717e75ecac

+++ PhzQtUI/src/lib/DialogLuminositySedGroup.cpp
@@ -20,7 +20,7 @@
 
 #include "PhzQtUI/SedGroupModel.h"
 
-#include <iostream>
+
 
 namespace Euclid {
 namespace PhzQtUI {
@@ -210,7 +210,7 @@ void DialogLuminositySedGroup::on_btn_cancel_clicked(){
 void DialogLuminositySedGroup::on_btn_save_clicked(){
   readNewGroups();
 
-  for (int i=0; i <m_groups.size();++i){
+  for (size_t i=0; i <m_groups.size();++i){
     auto& group_1 = m_groups[i];
 
     if (group_1.second.size()==0){
@@ -234,7 +234,7 @@ void DialogLuminositySedGroup::on_btn_save_clicked(){
 
 
 
-    for (int j=i+1; j <m_groups.size();++j){
+    for (size_t j=i+1; j <m_groups.size();++j){
        auto& group_2 = m_groups[j];
         if (group_1.first==group_2.first){
           QMessageBox::warning(this,
